requirements.txt
Django==5.2.8
django-allauth==0.57.0
Pillow==10.4.0
psycopg2-binary>=2.9.9
whitenoise==6.7.0
gunicorn==21.2.0
markdown==3.5.1
bleach==6.1.0
django-ratelimit==4.1.0
python-decouple==3.8
dj-database-url==2.1.0